Move dtos to its own file
