+++ services/worker/src/worker/job_runners/config/info.py
@@ -1,16 +1,11 @@
 import logging
-from typing import Any, Dict, TypedDict
 
 from libcommon.constants import PROCESSING_STEP_CONFIG_INFO_VERSION
 from libcommon.exceptions import PreviousStepFormatError
 from libcommon.simple_cache import get_previous_step_or_raise
 
+from worker.dtos import CompleteJobResult, ConfigInfoResponse
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
-from worker.utils import CompleteJobResult
-
-
-class ConfigInfoResponse(TypedDict):
-    dataset_info: Dict[str, Any]
 
 
 def compute_config_info_response(dataset: str, config: str) -> ConfigInfoResponse:
