+++ services/worker/src/worker/job_runners/config/opt_in_out_urls_count.py
@@ -13,8 +13,8 @@ from libcommon.simple_cache import (
     get_response,
 )
 
+from worker.dtos import JobResult, OptInOutUrlsCountResponse
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
-from worker.utils import JobResult, OptInOutUrlsCountResponse
 
 
 def compute_opt_in_out_urls_scan_response(dataset: str, config: str) -> Tuple[OptInOutUrlsCountResponse, float]:
