+++ services/worker/src/worker/job_runners/dataset/opt_in_out_urls_count.py
@@ -13,8 +13,8 @@ from libcommon.simple_cache import (
     get_response,
 )
 
+from worker.dtos import JobResult, OptInOutUrlsCountResponse
 from worker.job_runners.dataset.dataset_job_runner import DatasetJobRunner
-from worker.utils import JobResult, OptInOutUrlsCountResponse
 
 
 def compute_opt_in_out_urls_count_response(dataset: str) -> Tuple[OptInOutUrlsCountResponse, float]:
