+++ services/worker/src/worker/job_runners/split/first_rows_from_parquet.py
@@ -21,17 +21,15 @@ from libcommon.utils import JobInfo
 from libcommon.viewer_utils.features import get_cell_value
 
 from worker.config import AppConfig, FirstRowsConfig
-from worker.job_runners.split.split_job_runner import SplitJobRunner
-from worker.utils import (
+from worker.dtos import (
     CompleteJobResult,
     JobRunnerInfo,
     Row,
     RowItem,
     SplitFirstRowsResponse,
-    create_truncated_row_items,
-    get_json_size,
-    to_features_list,
 )
+from worker.job_runners.split.split_job_runner import SplitJobRunner
+from worker.utils import create_truncated_row_items, get_json_size, to_features_list
 
 
 def transform_rows(
