+++ services/worker/src/worker/job_runners/split/first_rows_from_streaming.py
@@ -33,12 +33,9 @@ from libcommon.utils import JobInfo
 from libcommon.viewer_utils.features import get_cell_value
 
 from worker.config import AppConfig, FirstRowsConfig
+from worker.dtos import CompleteJobResult, JobRunnerInfo, Row, SplitFirstRowsResponse
 from worker.job_runners.split.split_job_runner import SplitJobRunnerWithDatasetsCache
 from worker.utils import (
-    CompleteJobResult,
-    JobRunnerInfo,
-    Row,
-    SplitFirstRowsResponse,
     create_truncated_row_items,
     get_json_size,
     get_rows_or_raise,
