+++ services/worker/tests/job_runners/config/test_config_job_runner.py
@@ -9,8 +9,8 @@ from libcommon.processing_graph import ProcessingStep
 from libcommon.utils import Priority
 
 from worker.config import AppConfig
+from worker.dtos import CompleteJobResult
 from worker.job_runners.config.config_job_runner import ConfigJobRunner
-from worker.utils import CompleteJobResult
 
 
 class DummyConfigJobRunner(ConfigJobRunner):
