+++ services/worker/tests/job_runners/test__job_runner_with_datasets_cache.py
@@ -11,11 +11,11 @@ from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.utils import Priority
 
 from worker.config import AppConfig
+from worker.dtos import CompleteJobResult
 from worker.job_runners._job_runner_with_datasets_cache import (
     JobRunnerWithDatasetsCache,
 )
 from worker.resources import LibrariesResource
-from worker.utils import CompleteJobResult
 
 from ..fixtures.hub import get_default_config_split
