+++ services/worker/tests/test_loop.py
@@ -5,11 +5,11 @@ from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.utils import JobInfo
 
 from worker.config import AppConfig
+from worker.dtos import CompleteJobResult
 from worker.job_runner import JobRunner
 from worker.job_runner_factory import BaseJobRunnerFactory
 from worker.loop import Loop
 from worker.resources import LibrariesResource
-from worker.utils import CompleteJobResult
 
 
 class DummyJobRunner(JobRunner):
